requirements.txt
pyyaml==6.0.1
flet==0.23.2